+++ src/commonTest/kotlin/com/hoc081098/flowext/utils/BaseTest.kt
@@ -25,16 +25,22 @@
 package com.hoc081098.flowext.utils
 
 import kotlinx.coroutines.ExperimentalCoroutinesApi
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.collect
+import kotlinx.coroutines.flow.fold
+import kotlinx.coroutines.suspendCancellableCoroutine
 import kotlinx.coroutines.test.TestDispatcher
 import kotlinx.coroutines.test.TestResult
 import kotlinx.coroutines.test.TestScope
 import kotlinx.coroutines.test.UnconfinedTestDispatcher
+import kotlin.test.assertTrue
+import kotlin.test.fail
 
 abstract class BaseTest {
   @ExperimentalCoroutinesApi
   protected fun runTest(
     testDispatcher: TestDispatcher? = null,
-    testBody: suspend TestScope.() -> Unit
+    testBody: suspend TestScope.() -> Unit,
   ): TestResult {
     return kotlinx.coroutines.test.runTest(
       context = testDispatcher ?: UnconfinedTestDispatcher(name = "${this::class.simpleName}-dispatcher"),
@@ -42,3 +48,22 @@ abstract class BaseTest {
     )
   }
 }
+
+suspend inline fun hang(onCancellation: () -> Unit) {
+  try {
+    suspendCancellableCoroutine<Unit> { }
+  } finally {
+    onCancellation()
+  }
+}
+
+suspend inline fun <reified T : Throwable> assertFailsWith(flow: Flow<*>) {
+  try {
+    flow.collect()
+    fail("Should be unreached")
+  } catch (e: Throwable) {
+    assertTrue(e is T, "Expected exception ${T::class}, but had $e instead")
+  }
+}
+
+suspend fun Flow<Int>.sum() = fold(0) { acc, value -> acc + value }